Move sorter functions to util.go
